+++ com.unity.render-pipelines.universal/ShaderLibrary/ShaderQualityMedium.cs
@@ -5,7 +5,7 @@ namespace UnityEngine.Rendering.Universal
     {
         public static int reflectionProbe = 1;
         public static int blendReflectionProbe = 1;
-        public static int bumpScale = 0;
+        public static int bumpScale = 1;
         public static int fadeShadows = 1;
     }
 }