+++ com.unity.render-pipelines.universal/CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added the option to specify the maximum number of visible lights. If you set a value, lights are sorted based on their distance from the Camera.
 - Added the option to control the transparent layer separately in the Forward Renderer.
 - Added the option to specify shader quality which is used by the shaders to know which features to use.
-- Added Shadow fading wich uses the shadow quality
+- Added Shadow fading which uses the shadow quality
 
 ### Changed
 - Moved the icon that indicates the type of a Light 2D from the Inspector header to the Light Type field.